refactor: move props and state interfaces to component files
